Added a user config system
If a `user` lua module is provided (under `lua/user`) it will be used to
set:
- The LSP/Formatting servers to be installed via Mason and their
  configurations
- The colorscheme to use (only the ones provided by themer are supported
  for now)

If the user module is not provided or if the some config aren't included
within, default values will be used.

An example user config file is provided for now. It will probably move
to an example/doc folder later on.