feat: Allow user configuration
Provide a new user configuration system.
All configuration modules are now providing a `setup` function that has to be called
to configure the module.

---

If a `user` lua module is provided (under `lua/user`) it will be used to
set:
- The LSP/Formatting servers to be installed via Mason and their
  configurations
- The colorscheme to use (only the ones provided by themer are supported
  for now)

If the user module is not provided or if the some config aren't included
within, default values will be used.

An example user config file is provided for now. It will probably move
to an example/doc folder later on.

---

Renamed the lsp config module into language_support. This was reflected
on the user configuration keys.
This renamed module also takes care of treesitter configuration

The supported languages by treesitter can be configured via the user
configuration under the  `language_support.supported_languages` config.