Generalize user config
For each module, we now look for an existing user config key with the
same same. If the key is found, its contents are passed to the setup
function for that module. If it is not found, then nil is passed
instead.

This doesn't add any functionality as the parameters will be ignored by
modules not supporting them, but that simplifies the main config file
and should make adding/modifying/removing config modules esier in the
future